+++ lib/additionals/patches/query_patch.rb
@@ -50,7 +50,6 @@ module Additionals
           return unless available_filters[field]
 
           initialize_user_values_for_select2 field, values
-          true
         end
 
         def add_available_filter_with_additionals(field, options)